Add link to accessibillity statement
Add accessibility to the page footer via custom footer HTML option.

As the custom footer gets appended after the page footer which has a big bottom margin, we need to adjust it.
Custom footer HTML makes use of the .container class which is what the page uses, so we can benefit from content alignment