Rewrite headings on all pages
Makes the following changes, to introduce a
heading structure to all pages that:
- has a h1 for each page
- organises its headings in the right hierarchy

All pages now have a visually hidden h2 to
identify the footer section.

The home and history pages, which both have the
logo and subscribe to updates popup, have the logo
text rewritten to 'GOV.UK Notify'. This helps it
not clash with the home page h1 and makes it match
what it links to.

The home page has the most changes with:
- the introduction of a h1
- all content related to the current status being
  grouped and labelled with a h2
- the 'About' content being moved to the top and
  now forming a set of lead paragraphs
- all dates being reformatted to match other bits
  of Notify
- the 'Past incidents' section being renames
  'Recent incidents' to reflect it containing
  any that happened today

The history page has:
- the h4 at the top turned into a h1
- all h4s for each month turned into h2s

The incident page had mostly well-structured
headings so only has one added, to label the
section listing all affected components.

Co-authored-by: karlchillmaid <karl.chillmaid@digital.cabinet-office.gov.uk>