Re-run removing expand incidents on DOM changes
The status page JS changes the HTML of the list of
incidents-by-month when you click the pagination
buttons. This means any changes our JS makes to
the page HTML when it loads are replaced by the
original page HTML.

So the JS that removes the button to expand
incidents when there are more than 3 only works
on pages loaded by the browser not by JS.

This makes our JS also run when the status page JS
loads new incidents.