Make Camel-cased text sentence-case
Uses XPath to identify and target text nodes,
because most of the text we need to change is in a
text node with other sibling nodes of different
types. CSS doesn't allow us to target text nodes
so this is used to check the text node is where
expected and to avoid messing with the other
siblings.

But yeah: XPath. I know.