Add reformatting of links and headings
When the linting was added, it moved the JS from
custom-footer.html out into custom-footer.js.
Because this created a new file, changes added to
main in the meantime were missed from a rebase so
got lost.

This adds them back again.

The PR that originally added them is here:

https://github.com/alphagov/statuspage-customisations/pull/18