Fix XPath for footer link
The footer link is formatted nicely, so has its
contents on a new line, indented to show its place
the tag hierarchy. Unfortunately, this means the
first child node is a text node like '\n   ',
which means the current XPath matches that, not
the actual text of the link.

This changes it, to instead target the text node
after the ← character, which ignores surrounding
whitespace and is consistent across all pages.