Add testing and development dependencies
Cheerio to parse fetched source and modify
it.

Nunjucks to load the saved modified file
and serve it as template into which we load local
customisations.

[Webdriver.io](https://webdriver.io/) to run tests.
We use their local runner with Jasmine and spec
reporter.

[Concurently](https://www.npmjs.com/package/concurrently)
enables use to run dependent tasks such as server
and tests in one command instead of running them
into separate termibal windows.

Nodemon to watch for file changes and restart
the server.