Helpers functions: Load local code
We want to load the footer HMTL and Javascript
customisations as well as custom CSS they way
StatusPage does them.

We read all the files listed in the config array,
combine custom footer HTML and Javascript files
into one and then return a data object that
contains the above as well as the custom CSS.
This is then served into the template placeholders.