Helper functions: Load templates
When the server is starting it wants to load in
templates to serve. This helper function
checks if files exist before attempting to load them.

If they do, then we read the file and return its
contents.
If they don't, we run the function that generates
them from remote source. Once completed,
we then read them as normal.