Add incident descriptors
Instead of only relying on colours to identify
the severity of the incident, this change adds
a text descriptor.

On the history page we add a new container
under a link and use aria-describedby to couple
the link and the descriptor so that on tab,
link text plus descriptor text is announced.

On the incident page, we insert the descriptor
into the heading before the incident title.

In both cases, we map the classes elements are
given by StatusPage to the description text.

Positioning and looks were made with input from
Chris. We also decided to left align the incident
heading.