Update component list semantics
Component list on the homepage is just a
collection of divs, making it very difficult to
parse with a11y tools.

This change does two things
- changes semantics to an unordered list of
items
- adds visually-hidden descriptors for
component

To transform to the unordered list without a
visual change or writting more custom css,
exisitng element classes are attached to the
new markup.