Set up server entry point and npm script
This is mostly taken from Tom B.'s origginal work
on this with only minor modifications.

A simple http server that loads templates and
local customisation files then renders templates
with Nunjucks.

NPM script added that uses Nodemon and it's config
to watch for certain file changes and restart
the server.