Make pageStatus section handle variants
Because we had an unresolved incident today, we
saw that this state of the home page swaps the
.page-status component out for a
.unresolved-incidents one which also differs by
not having a h2 inside.

These changes alter the JS to deal with both
states, and the tests. The tests are also adapted
to deal with the fact that the
.unresolved-incidents component contains a
.incident-title as a child, which breaks our
previous understanding that those sections would
only appear in the incident list.