Move CSS code into src folder as SCSS
Replicates the order of the CSS in the original
`custom.css`

Adds a `npm run build:css` npm script which
produces `custom.css` in `dist`.

This adds some options to our nodemon config to
ignore the source .scss files so, at this stage,
we're leaving rebuilds of `custom.css` to the
user, only refreshing the server's copy of it when
`dist/custom.css` changes.

Includes changes to the stylelint config to allow
comments in the output CSS, to match the existing
`custom.css`.