Move JS code into src folder as .mjs
Moves out chunks of functionality and wraps them
in functions. These functions are then called at
the point the chunk existed in the original
`custom-footer.js`.

Adds a `npm run build:js` npm script which
produces `custom-footer.js` in `dist`.

Also adds a `npm run build:html` npm script which
copies `custom-footer.html` from the `src` to the
`dist` folder, and a `npm run build` to build all
static assets.

This adds some options to our nodemon config to
ignore the source .mjs files so, at this stage,
we're leaving rebuilds of `custom-footer.mjs` to
the user, only refreshing the server's copy of it
when `dist/custom-footer.mjs` changes.