Fix issues from changes to live /history page
The /history page has been changed so it loads
without any HTML content between the heading and
footer. The content is there but in JSON, in an
attribute value in the HTML. The statuspage JS
takes that and turns it into the HTML content,
which is then added to the page.

Because of this, our JS, which assumed the page
content would be available when the page loaded,
failed silently.

This changes our JS to watch the main content area
and run only when it has been filled by the
statuspage JS.

The webdriver.io testsuite we use is designed to
operate on DOM asynchronously by default, waiting
for all elements used by the tests until they are
in the DOM and available:

https://webdriver.io/docs/autowait

Because of this, we don't need to change our
tests for this, just our JS.