Refactor repo to support different statuspages
Introduces two things to allow this:
- a folder in the `./src` folder specific to the
  team, in which all the base SCSS, JS and HTML
  files are kept
- an `.env` file for environment variables, to
  make build and server code aware of which team
  is being used, their statuspage URL and the SHA
  for a real incident on their statuspage
- folders in the `./dist` folder for each team,
  which are automatically created if they don't
  exist

Most of the changes are made to accomodate this
new structure.

Bonus changes are that team-specific content is
also moved out of the generic code into the
team-specific files.

Add incident SHA to .env and constants

It's specific to each statuspage so needs to be
split out.